packages/flamenco-worker-python/requirements.txt
attrs==16.3.0
requests==2.12.4
